rules.abaplint.org refactoring
